Load NetworkManager as GlobalEntity
