+++ unity/Assets/0Refactored/GlobalEntities/GlobalEntityLoader/GlobalEntities.prefab
@@ -53,3 +53,5 @@ MonoBehaviour:
     type: 3}
   _soundManagerPrefab: {fileID: 4398862966192869678, guid: cd3a2bbefec729343ae52783e013e3c7,
     type: 3}
+  _networkManagerPrefab: {fileID: 2811340506612267987, guid: a5856923d4b332d43a49ce8ca34197b7,
+    type: 3}
