+++ unity/Assets/0Refactored/GlobalEntities/Network/scripts/NetworkStatusStation.cs
@@ -21,25 +21,25 @@ public class NetworkStatusStation : MonoBehaviour
     }
 
     private void UpdateNetworkInformation()
-    {
-        if (!MaroonNetworkManager.Instance.IsActive())
+    {        
+        if (!Maroon.NetworkManager.Instance.IsActive())
         {
             string msg = LanguageManager.Instance.GetString("ScreenHostJoin");
             DisplayMessage(msg);
             return;
         }
         
-        bool clientStarted = MaroonNetworkManager.Instance.mode == NetworkManagerMode.ClientOnly;
+        bool clientStarted = Maroon.NetworkManager.Instance.mode == NetworkManagerMode.ClientOnly;
         bool hostStarted = NetworkServer.active;
         bool listServerConnected = _ls.GetListServerStatus();
-        bool portsMapped = MaroonNetworkManager.Instance.PortsMapped;
+        bool portsMapped = Maroon.NetworkManager.Instance.PortsMapped;
         
         if (clientStarted)
         {
             DisplayOnlineInformation();
             statusLight.color = Color.green;
             statusLight.gameObject.SetActive(true);
-            MaroonNetworkManager.Instance.ConnectionStatus = ConnectionState.ClientOnline;
+            Maroon.NetworkManager.Instance.ConnectionStatus = ConnectionState.ClientOnline;
             return;
         }
 
@@ -48,13 +48,13 @@ public class NetworkStatusStation : MonoBehaviour
             statusLight.gameObject.SetActive(false);
             if (listServerConnected)
             {
-                MaroonNetworkManager.Instance.ConnectionStatus = ConnectionState.Offline;
+                Maroon.NetworkManager.Instance.ConnectionStatus = ConnectionState.Offline;
                 string msg = LanguageManager.Instance.GetString("ScreenHostJoin");
                 DisplayMessage(msg);
             }
             else
             {
-                MaroonNetworkManager.Instance.ConnectionStatus = ConnectionState.OfflineNoConnectionToListServer;
+                Maroon.NetworkManager.Instance.ConnectionStatus = ConnectionState.OfflineNoConnectionToListServer;
                 string msg = LanguageManager.Instance.GetString("ScreenNoListServerJoin");
                 DisplayMessage(msg);
             }
@@ -64,20 +64,20 @@ public class NetworkStatusStation : MonoBehaviour
             statusLight.gameObject.SetActive(true);
             if (listServerConnected && portsMapped)
             {
-                MaroonNetworkManager.Instance.ConnectionStatus = ConnectionState.HostOnline;
+                Maroon.NetworkManager.Instance.ConnectionStatus = ConnectionState.HostOnline;
                 DisplayOnlineInformation();
                 statusLight.color = Color.green;
             }
             else if (portsMapped) //List server not connected
             {
-                MaroonNetworkManager.Instance.ConnectionStatus = ConnectionState.HostOnlineNoConnectionToListServer;
+                Maroon.NetworkManager.Instance.ConnectionStatus = ConnectionState.HostOnlineNoConnectionToListServer;
                 string msg = LanguageManager.Instance.GetString("ScreenNoListServerHost");
                 DisplayMessage(msg);
                 statusLight.color = Color.red;
             }
             else //Ports not mapped
             {
-                MaroonNetworkManager.Instance.ConnectionStatus = ConnectionState.HostOnlinePortsNotMapped;
+                Maroon.NetworkManager.Instance.ConnectionStatus = ConnectionState.HostOnlinePortsNotMapped;
                 string msg = LanguageManager.Instance.GetString("ScreenNoPortsMapped");
                 DisplayMessage(msg);
                 statusLight.color = Color.yellow;
@@ -88,15 +88,15 @@ public class NetworkStatusStation : MonoBehaviour
     private void DisplayOnlineInformation()
     {
         string information = "";
-        information += LanguageManager.Instance.GetString("OnlineInformation1") + " <color=#800000>" + MaroonNetworkManager.Instance.PlayerName + "</color>!\n";
-        information += LanguageManager.Instance.GetString("OnlineInformation2") + " <color=#800000>" + MaroonNetworkManager.Instance.ServerName + "</color>.\n";
-        if (MaroonNetworkManager.Instance.IsInControl)
+        information += LanguageManager.Instance.GetString("OnlineInformation1") + " <color=#800000>" + Maroon.NetworkManager.Instance.PlayerName + "</color>!\n";
+        information += LanguageManager.Instance.GetString("OnlineInformation2") + " <color=#800000>" + Maroon.NetworkManager.Instance.ServerName + "</color>.\n";
+        if (Maroon.NetworkManager.Instance.IsInControl)
         {
             information += LanguageManager.Instance.GetString("OnlineInformation3");
         }
         else
         {
-            information += "<color=#800000>" + MaroonNetworkManager.Instance.ClientInControl + "</color> " + LanguageManager.Instance.GetString("OnlineInformation4");
+            information += "<color=#800000>" + Maroon.NetworkManager.Instance.ClientInControl + "</color> " + LanguageManager.Instance.GetString("OnlineInformation4");
         }
         DisplayMessage(information);
     }
