+++ unity/Assets/0Unused/oldMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 078180c6bcbdfbf46a0f85de8eb85b7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
